agent/requirements.txt
-i https://pypi.org/simple

boto3==1.40.62; python_version >= '3.9'
httpx==0.28.1; python_version >= '3.8'
livekit-agents[azure,assemblyai,deepgram,elevenlabs,images,openai,silero,turn-detector]==1.3.5; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
sentry-sdk==2.43.0; python_version >= '3.6'
tenacity==9.1.2; python_version >= '3.9'
